requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.2
django-cloudinary-storage==0.3.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
Markdown==3.10.2
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
